requirements.txt
django>=3.0.8
python-decouple==3.1
psycopg2
gunicorn
requests
pillow
unidecode
boto3
# https://github.com/darklow/django-suit/tarball/v2
djangorestframework
markdown
django-filter
python-dateutil
django-widget-tweaks
xmltodict
django-debug-toolbar
django-localflavor
pagseguro
djangorestframework-jwt
onesignal_sdk
django-storages
openpyxl
django-cms
onesignal
onesignal-client
flake8
isort
